change components
move to directory, and components to other files
